+++ x-pack/plugins/observability_solution/apm/server/routes/service_groups/get_service_group_alerts.ts
@@ -5,10 +5,11 @@
  * 2.0.
  */
 
-import { kqlQuery } from '@kbn/observability-plugin/server';
-import { ALERT_RULE_PRODUCER, ALERT_STATUS } from '@kbn/rule-data-utils';
+import { kqlQuery, termQuery, termsQuery } from '@kbn/observability-plugin/server';
+import { ALERT_RULE_PRODUCER, ALERT_STATUS, ALERT_STATUS_ACTIVE } from '@kbn/rule-data-utils';
 import { QueryDslQueryContainer } from '@elastic/elasticsearch/lib/api/types';
 import { Logger } from '@kbn/core/server';
+import { observabilityFeatureId } from '@kbn/observability-shared-plugin/common';
 import { ApmPluginRequestHandlerContext } from '../typings';
 import { SavedServiceGroup } from '../../../common/service_groups';
 import { ApmAlertsClient } from '../../lib/helpers/get_apm_alerts_client';
@@ -42,8 +43,8 @@ export async function getServiceGroupAlerts({
     query: {
       bool: {
         filter: [
-          { term: { [ALERT_RULE_PRODUCER]: 'apm' } },
-          { term: { [ALERT_STATUS]: 'active' } },
+          ...termsQuery(ALERT_RULE_PRODUCER, 'apm', observabilityFeatureId),
+          ...termQuery(ALERT_STATUS, ALERT_STATUS_ACTIVE),
         ],
       },
     },
