+++ x-pack/plugins/observability_solution/apm/server/routes/services/get_services/get_service_alerts.ts
@@ -5,13 +5,20 @@
  * 2.0.
  */
 
-import { kqlQuery, termQuery, rangeQuery, wildcardQuery } from '@kbn/observability-plugin/server';
+import {
+  kqlQuery,
+  termQuery,
+  rangeQuery,
+  wildcardQuery,
+  termsQuery,
+} from '@kbn/observability-plugin/server';
 import {
   ALERT_RULE_PRODUCER,
   ALERT_STATUS,
   ALERT_STATUS_ACTIVE,
   ALERT_UUID,
 } from '@kbn/rule-data-utils';
+import { observabilityFeatureId } from '@kbn/observability-shared-plugin/common';
 import { SERVICE_NAME } from '../../../../common/es_fields/apm';
 import { ServiceGroup } from '../../../../common/service_groups';
 import { ApmAlertsClient } from '../../../lib/helpers/get_apm_alerts_client';
@@ -51,7 +58,7 @@ export async function getServicesAlerts({
     query: {
       bool: {
         filter: [
-          ...termQuery(ALERT_RULE_PRODUCER, 'apm'),
+          ...termsQuery(ALERT_RULE_PRODUCER, 'apm', observabilityFeatureId),
           ...termQuery(ALERT_STATUS, ALERT_STATUS_ACTIVE),
           ...rangeQuery(start, end),
           ...kqlQuery(kuery),
