+++ x-pack/test/api_integration/deployment_agnostic/apis/observability/apm/alerts/transaction_error_rate.spec.ts
@@ -10,30 +10,41 @@ import { transactionErrorRateActionVariables } from '@kbn/apm-plugin/server/rout
 import { apm, timerange } from '@kbn/apm-synthtrace-client';
 import expect from '@kbn/expect';
 import { omit } from 'lodash';
-import { FtrProviderContext } from '../../common/ftr_provider_context';
+import type { ApmSynthtraceEsClient } from '@kbn/apm-synthtrace';
+import type { RoleCredentials, SupertestWithRoleScopeType } from '../../../../services';
+import type { DeploymentAgnosticFtrProviderContext } from '../../../../ftr_provider_context';
 import {
-  createApmRule,
   fetchServiceInventoryAlertCounts,
   fetchServiceTabAlertCount,
   ApmAlertFields,
   getIndexAction,
-  createIndexConnector,
-} from './helpers/alerting_api_helper';
-import { cleanupRuleAndAlertState } from './helpers/cleanup_rule_and_alert_state';
-import { waitForAlertsForRule } from './helpers/wait_for_alerts_for_rule';
-import { waitForActiveRule } from './helpers/wait_for_active_rule';
-import { waitForIndexConnectorResults } from './helpers/wait_for_index_connector_results';
+  APM_ACTION_VARIABLE_INDEX,
+  APM_ALERTS_INDEX,
+} from './helpers/alerting_helper';
 
-export default function ApiTest({ getService }: FtrProviderContext) {
-  const registry = getService('registry');
-  const supertest = getService('supertest');
-  const es = getService('es');
-  const logger = getService('log');
-  const apmApiClient = getService('apmApiClient');
-  const apmSynthtraceEsClient = getService('apmSynthtraceEsClient');
+export default function ApiTest({ getService }: DeploymentAgnosticFtrProviderContext) {
+  const roleScopedSupertest = getService('roleScopedSupertest');
+  const apmApiClient = getService('apmApi');
+  const synthtrace = getService('synthtrace');
+  const alertingApi = getService('alertingApi');
+  const samlAuth = getService('samlAuth');
 
-  registry.when('transaction error rate alert', { config: 'basic', archives: [] }, () => {
-    before(() => {
+  describe('transaction error rate alert', () => {
+    let apmSynthtraceEsClient: ApmSynthtraceEsClient;
+    let supertestViewerWithCookieCredentials: SupertestWithRoleScopeType;
+    let roleAuthc: RoleCredentials;
+
+    before(async () => {
+      supertestViewerWithCookieCredentials = await roleScopedSupertest.getSupertestWithRoleScope(
+        'viewer',
+        {
+          withInternalHeaders: true,
+          useCookieHeader: true,
+        }
+      );
+
+      roleAuthc = await samlAuth.createM2mApiKeyWithRoleScope('admin');
+
       const opbeansJava = apm
         .service({ name: 'opbeans-java', environment: 'production', agentName: 'java' })
         .instance('instance');
@@ -66,11 +77,15 @@ export default function ApiTest({ getService }: FtrProviderContext) {
               .success(),
           ];
         });
+
+      apmSynthtraceEsClient = await synthtrace.createApmSynthtraceEsClient();
       return apmSynthtraceEsClient.index(events);
     });
 
     after(async () => {
       await apmSynthtraceEsClient.clean();
+      await supertestViewerWithCookieCredentials.destroy();
+      await samlAuth.invalidateM2mApiKeyWithRoleScope(roleAuthc);
     });
 
     describe('create rule without kql query', () => {
@@ -79,16 +94,25 @@ export default function ApiTest({ getService }: FtrProviderContext) {
       let alerts: ApmAlertFields[];
 
       before(async () => {
-        actionId = await createIndexConnector({ supertest, name: 'Transation error rate' });
+        actionId = await alertingApi.createIndexConnector({
+          name: 'Transation error rate',
+          indexName: APM_ACTION_VARIABLE_INDEX,
+          roleAuthc,
+        });
+
         const indexAction = getIndexAction({
           actionId,
           actionVariables: transactionErrorRateActionVariables,
         });
 
-        const createdRule = await createApmRule({
-          supertest,
+        const createdRule = await alertingApi.createRule({
           ruleTypeId: ApmRuleType.TransactionErrorRate,
           name: 'Apm transaction error rate without kql query',
+          consumer: 'apm',
+          schedule: {
+            interval: '1m',
+          },
+          tags: ['apm'],
           params: {
             threshold: 40,
             windowSize: 5,
@@ -104,17 +128,33 @@ export default function ApiTest({ getService }: FtrProviderContext) {
             ],
           },
           actions: [indexAction],
+          roleAuthc,
         });
         ruleId = createdRule.id;
-        alerts = await waitForAlertsForRule({ es, ruleId });
+        alerts = (
+          await alertingApi.waitForDocumentInIndex({
+            indexName: APM_ALERTS_INDEX,
+            ruleId,
+          })
+        ).hits.hits.map((hit) => hit._source) as ApmAlertFields[];
       });
 
-      after(async () => {
-        await cleanupRuleAndAlertState({ es, supertest, logger });
-      });
+      after(() =>
+        alertingApi.cleanUpAlerts({
+          roleAuthc,
+          ruleId,
+          alertIndexName: APM_ALERTS_INDEX,
+          connectorIndexName: APM_ACTION_VARIABLE_INDEX,
+          consumer: 'apm',
+        })
+      );
 
       it('checks if rule is active', async () => {
-        const ruleStatus = await waitForActiveRule({ ruleId, supertest });
+        const ruleStatus = await alertingApi.waitForRuleStatus({
+          ruleId,
+          roleAuthc,
+          expectedStatus: 'active',
+        });
         expect(ruleStatus).to.be('active');
       });
 
@@ -122,7 +162,11 @@ export default function ApiTest({ getService }: FtrProviderContext) {
         let results: Array<Record<string, string>>;
 
         before(async () => {
-          results = await waitForIndexConnectorResults({ es });
+          results = results = (
+            await alertingApi.waitForDocumentInIndex({
+              indexName: APM_ACTION_VARIABLE_INDEX,
+            })
+          ).hits.hits.map((hit) => hit._source) as Array<Record<string, string>>;
         });
 
         it('has the right keys', async () => {
@@ -142,7 +186,7 @@ export default function ApiTest({ getService }: FtrProviderContext) {
         });
 
         it('has the right values', () => {
-          expect(omit(results[0], 'alertDetailsUrl')).to.eql({
+          expect(omit(results[0], 'alertDetailsUrl', 'viewInAppUrl')).to.eql({
             environment: 'production',
             interval: '5 mins',
             reason:
@@ -152,9 +196,13 @@ export default function ApiTest({ getService }: FtrProviderContext) {
             threshold: '40',
             transactionType: 'request',
             triggerValue: '50',
-            viewInAppUrl:
-              'http://mockedPublicBaseUrl/app/apm/services/opbeans-java?transactionType=request&environment=production',
           });
+
+          const url = new URL(results[0].viewInAppUrl);
+
+          expect(url.pathname).to.equal('/app/apm/services/opbeans-java');
+          expect(url.searchParams.get('transactionType')).to.equal('request');
+          expect(url.searchParams.get('environment')).to.equal('production');
         });
       });
 
@@ -201,10 +249,14 @@ export default function ApiTest({ getService }: FtrProviderContext) {
       let alerts: ApmAlertFields[];
 
       beforeEach(async () => {
-        const createdRule = await createApmRule({
-          supertest,
+        const createdRule = await alertingApi.createRule({
           ruleTypeId: ApmRuleType.TransactionErrorRate,
           name: 'Apm transaction error rate without kql query',
+          consumer: 'apm',
+          schedule: {
+            interval: '1m',
+          },
+          tags: ['apm'],
           params: {
             threshold: 40,
             windowSize: 5,
@@ -227,14 +279,26 @@ export default function ApiTest({ getService }: FtrProviderContext) {
             ],
           },
           actions: [],
+          roleAuthc,
         });
         ruleId = createdRule.id;
-        alerts = await waitForAlertsForRule({ es, ruleId });
+        alerts = (
+          await alertingApi.waitForDocumentInIndex({
+            indexName: APM_ALERTS_INDEX,
+            ruleId,
+          })
+        ).hits.hits.map((hit) => hit._source) as ApmAlertFields[];
       });
 
-      afterEach(async () => {
-        await cleanupRuleAndAlertState({ es, supertest, logger });
-      });
+      afterEach(() =>
+        alertingApi.cleanUpAlerts({
+          roleAuthc,
+          ruleId,
+          alertIndexName: APM_ALERTS_INDEX,
+          connectorIndexName: APM_ACTION_VARIABLE_INDEX,
+          consumer: 'apm',
+        })
+      );
 
       it('indexes alert document with all group-by fields', async () => {
         expect(alerts[0]).property('service.name', 'opbeans-node');
