+++ x-pack/test/api_integration/deployment_agnostic/services/alerting_api.ts
@@ -11,8 +11,9 @@ import type {
 } from '@elastic/elasticsearch/lib/api/typesWithBodyKey';
 import { MetricThresholdParams } from '@kbn/infra-plugin/common/alerting/metrics';
 import { ThresholdParams } from '@kbn/observability-plugin/common/custom_threshold_rule/types';
+import { ApmRuleParamsType } from '@kbn/apm-plugin/common/rules/schema';
 import { RoleCredentials } from '@kbn/ftr-common-functional-services';
-import type { Client } from '@elastic/elasticsearch';
+import { errors, type Client } from '@elastic/elasticsearch';
 import type { TryWithRetriesOptions } from '@kbn/ftr-common-functional-services';
 import { v4 as uuidv4 } from 'uuid';
 import moment from 'moment';
@@ -613,17 +614,6 @@ export function AlertingApiProvider({ getService }: DeploymentAgnosticFtrProvide
       return body;
     },
 
-    async findRuleById(roleAuthc: RoleCredentials, ruleId: string) {
-      if (!ruleId) {
-        throw new Error(`'ruleId' is undefined`);
-      }
-      const response = await supertestWithoutAuth
-        .get(`/api/alerting/rule/${ruleId}`)
-        .set(samlAuth.getInternalRequestHeader())
-        .set(roleAuthc.apiKeyHeader);
-      return response.body || {};
-    },
-
     waiting: {
       async waitForDocumentInIndex({
         esClient,
@@ -948,7 +938,6 @@ export function AlertingApiProvider({ getService }: DeploymentAgnosticFtrProvide
 
   return {
     helpers,
-
     async waitForRuleStatus({
       ruleId,
       expectedStatus,
@@ -976,14 +965,27 @@ export function AlertingApiProvider({ getService }: DeploymentAgnosticFtrProvide
     async waitForDocumentInIndex<T>({
       indexName,
       docCountTarget = 1,
+      ruleId,
     }: {
       indexName: string;
       docCountTarget?: number;
+      ruleId?: string;
     }): Promise<SearchResponse<T, Record<string, AggregationsAggregate>>> {
       return await retry.tryForTime(retryTimeout, async () => {
         const response = await es.search<T>({
           index: indexName,
           rest_total_hits_as_int: true,
+          ...(ruleId
+            ? {
+                body: {
+                  query: {
+                    term: {
+                      'kibana.alert.rule.uuid': ruleId,
+                    },
+                  },
+                },
+              }
+            : {}),
         });
         logger.debug(`Found ${response.hits.total} docs, looking for at least ${docCountTarget}.`);
 
@@ -1064,7 +1066,15 @@ export function AlertingApiProvider({ getService }: DeploymentAgnosticFtrProvide
     }: {
       ruleTypeId: string;
       name: string;
-      params: MetricThresholdParams | ThresholdParams | SloBurnRateRuleParams;
+      params:
+        | CreateEsQueryRuleParams
+        | MetricThresholdParams
+        | ThresholdParams
+        | SloBurnRateRuleParams
+        | ApmRuleParamsType['apm.anomaly']
+        | ApmRuleParamsType['apm.error_rate']
+        | ApmRuleParamsType['apm.transaction_duration']
+        | ApmRuleParamsType['apm.transaction_error_rate'];
       actions?: any[];
       tags?: any[];
       schedule?: { interval: string };
@@ -1095,6 +1105,104 @@ export function AlertingApiProvider({ getService }: DeploymentAgnosticFtrProvide
         .set(roleAuthc.apiKeyHeader)
         .set(samlAuth.getInternalRequestHeader());
       return response.body.data.find((obj: any) => obj.id === ruleId);
+    },
+
+    async searchRules(roleAuthc: RoleCredentials, filter: string) {
+      return supertestWithoutAuth
+        .get('/api/alerting/rules/_find')
+        .query({ filter })
+        .set(roleAuthc.apiKeyHeader)
+        .set(samlAuth.getInternalRequestHeader());
+    },
+
+    async deleteRuleById({ roleAuthc, ruleId }: { roleAuthc: RoleCredentials; ruleId: string }) {
+      return supertestWithoutAuth
+        .delete(`/api/alerting/rule/${ruleId}`)
+        .set(roleAuthc.apiKeyHeader)
+        .set(samlAuth.getInternalRequestHeader());
+    },
+
+    async deleteRules({ roleAuthc, filter }: { roleAuthc: RoleCredentials; filter: string }) {
+      const response = await this.searchRules(roleAuthc, filter);
+      return Promise.all(
+        response.body.data.map((rule: any) => this.deleteRuleById({ roleAuthc, ruleId: rule.id }))
+      );
+    },
+
+    async deleteAllActionConnectors({ roleAuthc }: { roleAuthc: RoleCredentials }): Promise<any> {
+      const res = await supertestWithoutAuth
+        .get(`/api/actions/connectors`)
+        .set(roleAuthc.apiKeyHeader)
+        .set(samlAuth.getInternalRequestHeader());
+
+      const body = res.body as Array<{ id: string; connector_type_id: string; name: string }>;
+      return Promise.all(
+        body.map(({ id }) => {
+          return this.deleteActionConnector({
+            roleAuthc,
+            actionId: id,
+          });
+        })
+      );
+    },
+
+    async deleteActionConnector({
+      roleAuthc,
+      actionId,
+    }: {
+      roleAuthc: RoleCredentials;
+      actionId: string;
+    }) {
+      return supertestWithoutAuth
+        .delete(`/api/actions/connector/${actionId}`)
+        .set(roleAuthc.apiKeyHeader)
+        .set(samlAuth.getInternalRequestHeader());
+    },
+
+    async cleanUpAlerts({
+      roleAuthc,
+      ruleId,
+      consumer,
+      alertIndexName,
+      connectorIndexName,
+    }: {
+      roleAuthc: RoleCredentials;
+      ruleId: string;
+      consumer?: string;
+      alertIndexName?: string;
+      connectorIndexName?: string;
+    }) {
+      return Promise.allSettled([
+        // Delete the rule by ID
+        this.deleteRuleById({ roleAuthc, ruleId }),
+        // Delete all documents in the alert index if specified
+        alertIndexName
+          ? es.deleteByQuery({
+              index: alertIndexName,
+              conflicts: 'proceed',
+              query: { match_all: {} },
+            })
+          : Promise.resolve(),
+        // Delete event logs for the specified consumer if provided
+        consumer
+          ? es.deleteByQuery({
+              index: '.kibana-event-log-*',
+              query: { term: { 'kibana.alert.rule.consumer': consumer } },
+            })
+          : Promise.resolve(),
+        // Delete connector index if provided
+        connectorIndexName
+          ? es.indices.delete({ index: connectorIndexName }).catch((e) => {
+              if (e instanceof errors.ResponseError && e.statusCode === 404) {
+                return;
+              }
+
+              throw e;
+            })
+          : Promise.resolve(),
+        // Delete all action connectors
+        this.deleteAllActionConnectors({ roleAuthc }),
+      ]);
     },
   };
 }
