+++ x-pack/test/apm_api_integration/tests/alerts/anomaly_alert.spec.ts
@@ -12,12 +12,12 @@ import { apm, timerange } from '@kbn/apm-synthtrace-client';
 import expect from '@kbn/expect';
 import { range } from 'lodash';
 import { ML_ANOMALY_SEVERITY } from '@kbn/ml-anomaly-utils/anomaly_severity';
+import { waitForAlertsForRule } from './helpers/wait_for_alerts_for_rule';
+import { waitForActiveRule } from './helpers/wait_for_active_rule';
+import { createApmRule } from './helpers/alerting_api_helper';
+import { cleanupRuleAndAlertState } from './helpers/cleanup_rule_and_alert_state';
 import { FtrProviderContext } from '../../common/ftr_provider_context';
 import { createAndRunApmMlJobs } from '../../common/utils/create_and_run_apm_ml_jobs';
-import { createApmRule } from './helpers/alerting_api_helper';
-import { waitForActiveRule } from './helpers/wait_for_active_rule';
-import { waitForAlertsForRule } from './helpers/wait_for_alerts_for_rule';
-import { cleanupRuleAndAlertState } from './helpers/cleanup_rule_and_alert_state';
 
 export default function ApiTest({ getService }: FtrProviderContext) {
   const registry = getService('registry');
