+++ x-pack/test/apm_api_integration/tests/alerts/helpers/wait_for_alerts_for_rule.ts
@@ -11,7 +11,10 @@ import type {
   SearchResponse,
 } from '@elastic/elasticsearch/lib/api/typesWithBodyKey';
 import pRetry from 'p-retry';
-import { ApmAlertFields, APM_ALERTS_INDEX } from './alerting_api_helper';
+import {
+  APM_ALERTS_INDEX,
+  ApmAlertFields,
+} from '../../../../api_integration/deployment_agnostic/apis/observability/apm/alerts/helpers/alerting_helper';
 
 async function getAlertByRuleId({ es, ruleId }: { es: Client; ruleId: string }) {
   const response = (await es.search({
