+++ x-pack/test/apm_api_integration/tests/services/service_alerts.spec.ts
@@ -8,10 +8,10 @@ import expect from '@kbn/expect';
 import { AggregationType } from '@kbn/apm-plugin/common/rules/apm_rule_types';
 import { ApmRuleType } from '@kbn/rule-data-utils';
 import { apm, timerange } from '@kbn/apm-synthtrace-client';
+import { waitForAlertsForRule } from '../alerts/helpers/wait_for_alerts_for_rule';
 import { FtrProviderContext } from '../../common/ftr_provider_context';
 import { createApmRule, runRuleSoon, ApmAlertFields } from '../alerts/helpers/alerting_api_helper';
 import { waitForActiveRule } from '../alerts/helpers/wait_for_active_rule';
-import { waitForAlertsForRule } from '../alerts/helpers/wait_for_alerts_for_rule';
 import { cleanupRuleAndAlertState } from '../alerts/helpers/cleanup_rule_and_alert_state';
 
 export default function ServiceAlerts({ getService }: FtrProviderContext) {
